requirements-dev.txt
black==23.11.0
mypy==1.6.1
ruff==0.1.5
pandas-stubs==2.1.1.230928
